pyproject.toml
[tool.poetry]
name = "automationexercise-autotests"
version = "0.1.0"
description = ""
authors = ["Eugene Evdokimenko <eugene.qa94@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
selene = "2.0.0rc9"
allure-pytest = "^2.13.5"
selenium = ">=4.21.0,<4.22.0"
python-dotenv = ">=1.0.1,<1.1.0"
requests = "^2.32.3"
appium-python-client = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
